# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for your monorepo services",
        "title": "Your Monorepo API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/api/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves paginated products optionally filtered by search term",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get list of products",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default: 10)",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term to filter products by name or description",
                        "name": "searchTerm",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new product with the provided information. Category must exist and be valid. Request body should only contain category_id, not the full category object.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Create a new product",
                "parameters": [
                    {
                        "description": "Product data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ProductCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "schema": {
                            "$ref": "#/definitions/model.Product"
                        }
                    },
                    "400": {
                        "description": "Validation error (invalid input or category not found)",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/api/products/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a specific product by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get product by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID (UUID format)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing product with new information. Category must exist and be valid. Request body should only contain category_id, not the full category object.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Update a product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID (UUID format)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated product data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ProductUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "schema": {
                            "$ref": "#/definitions/model.Product"
                        }
                    },
                    "400": {
                        "description": "Validation error (invalid input or category not found)",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a product by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Delete a product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID (UUID format)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ProductCreateRequest": {
            "type": "object",
            "required": [
                "category_id",
                "code",
                "content_per_large_unit",
                "large_unit",
                "name",
                "purchase_price",
                "selling_price",
                "small_unit"
            ],
            "properties": {
                "category_id": {
                    "description": "Foreign key to category",
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "code": {
                    "description": "Product code or SKU",
                    "type": "string",
                    "example": "PROD001"
                },
                "content_per_large_unit": {
                    "description": "e.g., 12 pieces per box",
                    "type": "integer",
                    "minimum": 1,
                    "example": 12
                },
                "indication": {
                    "description": "Description or usage",
                    "type": "string",
                    "example": "High-performance laptop for professionals"
                },
                "large_unit": {
                    "description": "e.g., box, pack",
                    "type": "string",
                    "example": "box"
                },
                "name": {
                    "description": "Product name",
                    "type": "string",
                    "example": "Laptop Dell XPS 13"
                },
                "purchase_price": {
                    "description": "Cost price",
                    "type": "number",
                    "minimum": 0,
                    "example": 500
                },
                "selling_price": {
                    "description": "Sale price",
                    "type": "number",
                    "minimum": 0,
                    "example": 750
                },
                "small_unit": {
                    "description": "e.g., piece, tablet",
                    "type": "string",
                    "example": "piece"
                }
            }
        },
        "dto.ProductUpdateRequest": {
            "type": "object",
            "required": [
                "category_id",
                "code",
                "content_per_large_unit",
                "large_unit",
                "name",
                "purchase_price",
                "selling_price",
                "small_unit"
            ],
            "properties": {
                "category_id": {
                    "description": "Foreign key to category",
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "code": {
                    "description": "Product code or SKU",
                    "type": "string",
                    "example": "PROD001"
                },
                "content_per_large_unit": {
                    "description": "e.g., 12 pieces per box",
                    "type": "integer",
                    "minimum": 1,
                    "example": 12
                },
                "indication": {
                    "description": "Description or usage",
                    "type": "string",
                    "example": "High-performance laptop for professionals"
                },
                "large_unit": {
                    "description": "e.g., box, pack",
                    "type": "string",
                    "example": "box"
                },
                "name": {
                    "description": "Product name",
                    "type": "string",
                    "example": "Laptop Dell XPS 13"
                },
                "purchase_price": {
                    "description": "Cost price",
                    "type": "number",
                    "minimum": 0,
                    "example": 500
                },
                "selling_price": {
                    "description": "Sale price",
                    "type": "number",
                    "minimum": 0,
                    "example": 750
                },
                "small_unit": {
                    "description": "e.g., piece, tablet",
                    "type": "string",
                    "example": "piece"
                }
            }
        },
...
        "model.CategoryProduct": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Timestamp when created",
                    "type": "string"
                },
                "id": {
                    "description": "Unique identifier",
                    "type": "string"
                },
                "name": {
                    "description": "Category name",
                    "type": "string"
                },
                "parent_id": {
                    "description": "Parent category ID",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Timestamp when updated",
                    "type": "string"
                }
            }
        },
...
        "model.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/model.CategoryProduct"
                },
                "category_id": {
                    "description": "Foreign key",
                    "type": "string"
                },
                "code": {
                    "description": "Product code or SKU",
                    "type": "string"
                },
                "content_per_large_unit": {
                    "description": "e.g., 12 pieces per box",
                    "type": "integer"
                },
                "created_at": {
                    "description": "Timestamp when created",
                    "type": "string"
                },
                "id": {
                    "description": "Unique identifier",
                    "type": "string"
                },
                "indication": {
                    "description": "Description or usage",
                    "type": "string"
                },
                "large_unit": {
                    "description": "e.g., box, pack",
                    "type": "string"
                },
                "name": {
                    "description": "Product name",
                    "type": "string"
                },
                "purchase_price": {
                    "description": "Cost price",
                    "type": "number"
                },
                "selling_price": {
                    "description": "Sale price",
                    "type": "number"
                },
                "small_unit": {
                    "description": "e.g., piece, tablet",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Timestamp when updated",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}